requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
mypy==0.812
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0

pytest==4.6.5
pytest-runner==5.1
pytest-mock==3.6.1
featureflags~=0.1.0
setuptools~=54.2.0
httpx~=0.16.1
attrs~=20.3.0

pyjwt==2.0.1